+++ assets/js/vue-apps/form-components/address-lookup.vue
@@ -107,8 +107,8 @@ export default {
             return (
                 this.candidates.length === 0 &&
                 this.postcode &&
-                (this.currentState !== states.NotRequired &&
-                    this.currentState !== states.EnteringManually)
+                this.currentState !== states.NotRequired &&
+                    this.currentState !== states.EnteringManually
             );
         },
         getAddressFromId(udprn) {
