+++ controllers/apply/form-router/delete.js
