+++ controllers/apply/standard-proposal/index.js
@@ -2,7 +2,7 @@
 const path = require('path');
 const features = require('config').get('features');
 
-const { initFormRouter } = require('../form-router-next');
+const { initFormRouter } = require('../form-router');
 
 const formBuilder = require('./form');
 const confirmationBuilder = require('./confirmation');
